Separate Types/Funcs from main ReduxDucks
